+++ CHANGELOG.md
@@ -1,7 +1,11 @@
+## Upcoming
+
+* feat: add `SettingsTile.checkboxTile` support.
+
 ## 1.0.0+1
 
-* Fix broken image
+* Fix broken image.
 
 ## 1.0.0
 
-* First stable release
+* First stable release.